added bruteforce method for FS-possibilities
now the movessequence function returns a list of possibilities how to solve the problem with the FS algorithm.
If pegheightdicts with equal heights in different order are removed, there are no equal possibilities.
However, if they are not removed, there are more possibilities.
If one applies the removeequals function to the moveslist,
it removes all equal movelists and returns a list of THobjects with the history according to the moves.
Not removing the equal-height pegheightdicts but in the end removing equal movesequences leads to more different movesequences than just removing pegheightdicts.

Also added the bruteforceadjustedUpsilon function to visualize so now one can create a table with its values.
Removed the comparehs function, because the removeequals with option show = True yields the same functionality.