improved visualization, answered mathematical question
added showmoveslist and showmoveslists to visualization.
showmoveslist shows the configurations of a moveslist one after another
showmoveslists takes multiple moveslists, shows all configurations with the same number of moves next to each other
and inserts a newpage if the number of moves increases.

Answered the question why adjustedUpsilon and bruteforceadjustedUpsilon return different results.